+++ packages/sdk-react-ui/CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0]
+### Added
+- feat: implementing internationalization via i18next package ([#403](https://github.com/MetaMask/metamask-sdk/pull/403))
+
 ## [0.8.0]
-### Uncategorized
+### Added
 - feat: adds contact form to npm packages ([#382](https://github.com/MetaMask/metamask-sdk/pull/382))
 
 ## [0.7.0]
@@ -28,7 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [feat] initial beta released
 
-[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/MetaMask/metamask-sdk/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/MetaMask/metamask-sdk/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/MetaMask/metamask-sdk/compare/v0.6.2...v0.7.0
 [0.6.2]: https://github.com/MetaMask/metamask-sdk/compare/v0.6.1...v0.6.2